src_py/requirements.txt
grpcio
grpcio-tools
openai
python-dotenv
httpx
sympy
